feat: revamp the chow library
* create a chow package for the functions that handle stuff
* add tests (even though they're dummy at this point)
* make use of go modules
