+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/ichtrojan/chow
+
+go 1.12
+
+require github.com/urfave/cli v1.22.2
